// frontend/website/src/lib/types/notification.api.ts
export interface paths {
    "/health": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["health"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/internal/delete/{uuid}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post?: never;
        delete: operations["delete_handler"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/internal/global_message": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["global_message"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/internal/invalidate/{uuid}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["invalidate_user"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/internal/role_message": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["role_message"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/internal/user_message/{uuid}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["internal_user_message"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/subscribe": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["push_api_subscribe"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/user_message/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["user_message"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/ws/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["notify"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        AuthenticatedUser: {
            role: components["schemas"]["UserRole"];
            /** Format: uuid */
            session_id: string;
            username: string;
            /** Format: uuid */
            uuid: string;
        };
        Claims: {
            /** Format: date-time */
            exp: string;
            /** Format: date-time */
            iat: string;
            public_key?: string | null;
            roles: components["schemas"]["HashMap"];
            /** Format: uuid */
            session_id: string;
            /** Format: uuid */
            sub: string;
            username: string;
        };
        HashMap: {
            [key: string]: "devin" | "owen" | "mrd" | "trusted" | "user";
        };
        /** @enum {string} */
        Namespaces: "notification" | "nanopass" | "gradegetter" | "smalltalk_keysync" | "smalltalk_notes";
        RoleMessage: {
            message: string;
            target_role: components["schemas"]["UserRole"];
        };
        SendNotification: {
            content: string;
            recipient: string;
        };
        SubscribeRequest: {
            endpoint: string;
            keys: components["schemas"]["SubscriptionKeys"];
        };
        SubscriptionKeys: {
            auth: string;
            p256dh: string;
        };
        /** @enum {string} */
        UserRole: "devin" | "owen" | "mrd" | "trusted" | "user";
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}